Fix(Profile):Fix the user profile for avatar delete

+++ src/main/java/com/luxtix/eventManagementWebsite/users/service/impl/UserServiceImpl.java
@@ -106,10 +106,13 @@ public class UserServiceImpl implements UserService {
     public ProfileResponseDto updateProfile(String email, ProfileRequestDto profileRequestDto) {
         Users userData = userRepository.findByEmail(email).orElseThrow(() -> new DataNotFoundException("User not found"));
         if(profileRequestDto.getAvatar() != null){
-            try {
-                cloudinaryService.deleteImage(userData.getAvatar());
-            } catch (IOException e) {
-                throw new RuntimeException(e);
+            if(userData.getAvatar() != null)
+            {
+                try {
+                    cloudinaryService.deleteImage(userData.getAvatar());
+                } catch (IOException e) {
+                    throw new RuntimeException(e);
+                }
             }
             userData.setAvatar(cloudinaryService.uploadFile(profileRequestDto.getAvatar(),"folder_luxtix"));
         }
@@ -124,6 +127,7 @@ public class UserServiceImpl implements UserService {
         ProfileResponseDto data = new ProfileResponseDto();
         data.setEmail(userData.getEmail());
         data.setDisplayName(userData.getFullname());
+        System.out.println(userData.getAvatar());
         data.setAvatar(cloudinaryService.generateUrl(userData.getAvatar()));
         data.setPhoneNumber(userData.getPhoneNumber());
         data.setDisplayName(userData.getFullname());
